Change response message for client and vet info change so it would come as JSON, not in exception, rename Account test files
